Refactor package structure and update imports
Reorganized the SDK by moving modules under the zarinpal namespace, renaming 'resources' to 'models', and updating all relevant imports in examples, tests, and source files. This improves code organization and aligns import paths for better maintainability.
