+++ src/zarinpal-py-sdk/zarinpal/models/Fee.py
@@ -1,5 +1,5 @@
-from zarinpal import ZarinPal
-from utils.Validator import Validator
+from zarinpal.service import ZarinPal
+from zarinpal.utils import Validator
 
 class Fee:
     def __init__(self, zarinpal: ZarinPal):
