+++ src/zarinpal-py-sdk/zarinpal/models/Payments.py
@@ -1,5 +1,5 @@
 from typing import Dict, Any
-from zarinpal import ZarinPal
+from zarinpal.service import ZarinPal
 
 
 class Payments:
@@ -16,7 +16,7 @@ class Payments:
         self.endpoint = '/pg/v4/payment/request.json'
 
     def create(self, data: Dict[str, Any]) -> Dict[str, Any]:
-        from utils.Validator import Validator
+        from zarinpal.utils import Validator
         """
         Create a new payment request.
         :param data: A dictionary containing payment details:
